requirements.txt
customtkinter
opencv-python
mediapipe==0.10.11
pillow
